Split spectral WGPU shaders from signal ops
