+++ crates/runmat-accelerate/src/backend/wgpu/shaders/mod.rs
@@ -31,6 +31,7 @@ pub mod reduction;
 pub mod repmat;
 pub mod scan;
 pub mod scatter;
+pub mod signal;
 pub mod solve;
 pub mod stochastic_evolution;
 pub mod sub2ind;
